+++ AlchemistCode/Cards/Common/Grist.cs
@@ -1,0 +1,28 @@
+using System.Linq;
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.CardSelection;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+
+namespace Alchemist.AlchemistCode.Cards.Common;
+
+public class Grist : AlchemistCard
+{
+    public Grist() : base(1, CardType.Attack, CardRarity.Common, TargetType.AnyEnemy)
+    {
+        WithDamage(11, 3);
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_rock_shatter"),
+            tmpSfx: "heavy_attack.mp3").Execute(choiceContext);
+        var picked = (await CardSelectCmd.FromHandForDiscard(
+            choiceContext, Owner,
+            new CardSelectorPrefs(CardSelectorPrefs.DiscardSelectionPrompt, 1),
+            null, this)).FirstOrDefault();
+        if (picked != null)
+            await CardCmd.Discard(choiceContext, picked);
+    }
+}