+++ AlchemistCode/Cards/Uncommon/Ballast.cs
@@ -1,0 +1,21 @@
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+
+namespace Alchemist.AlchemistCode.Cards.Uncommon;
+
+public class Ballast : AlchemistCard
+{
+    public Ballast() : base(1, CardType.Attack, CardRarity.Uncommon, TargetType.AnyEnemy)
+    {
+        WithDamage(7, 3);
+        WithBlock(7, 3);
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_rock_shatter"),
+            tmpSfx: "heavy_attack.mp3").Execute(choiceContext);
+        await CommonActions.CardBlock(this, play);
+    }
+}