+++ AlchemistCode/Cards/Uncommon/Transmute.cs
@@ -1,13 +1,7 @@
-using Alchemist.AlchemistCode.Compat;
-using Alchemist.AlchemistCode.Powers;
+using Alchemist.AlchemistCode.Commands;
 using BaseLib.Utils;
-using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
-using MegaCrit.Sts2.Core.Entities.Potions;
-using MegaCrit.Sts2.Core.Factories;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
-using MegaCrit.Sts2.Core.Models.Powers;
-using MegaCrit.Sts2.Core.Models;
 
 namespace Alchemist.AlchemistCode.Cards.Uncommon;
 
@@ -16,18 +10,10 @@ public class Transmute : AlchemistCard
     public Transmute() : base(0, CardType.Skill, CardRarity.Uncommon, TargetType.Self)
     {
         WithKeyword(CardKeyword.Exhaust);
-        WithCards(1, 1);
-        WithTip(typeof(PoisonPower));
-        WithTip(typeof(StrengthPower));
+        WithVar("Cards", 2, 0);
+        WithTips(_ => Infusion.InfuseTips());
     }
 
-    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
-    {
-
-        var poison = Owner.Creature.GetPowerAmount<PoisonPower>();
-        if (poison > 0)
-            await PowerCmd.Apply<TransmuteStrengthPower>(choiceContext, Owner.Creature, poison, Owner.Creature, this);
-
-        await CommonActions.Draw(this, choiceContext);
-    }
+    protected override Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play) =>
+        Infusion.InfuseChosen(choiceContext, this, PileType.Hand, DynamicVars["Cards"].IntValue);
 }